# api/openapi.yaml
openapi: 3.0.0
info:
  title: Pizza API
  description: An API for managing pizzas and toppings
  version: 1.0
paths:
...
  /pizzas/{id}:
    get:
      summary: Get a pizza by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              example: {}
        '404':
          description: Pizza not found
    patch:
      summary: Update a pizza by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              example: {}
    delete:
      summary: Delete a pizza by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
...